[Java] Consolidate tagged and min flow control strategies.
