+++ aeron-driver/src/main/java/io/aeron/driver/TaggedMulticastFlowControl.java
@@ -18,14 +18,10 @@ package io.aeron.driver;
 import io.aeron.CommonContext;
 import io.aeron.driver.media.UdpChannel;
 import io.aeron.protocol.StatusMessageFlyweight;
-import org.agrona.AsciiEncoding;
 import org.agrona.BitUtil;
-import org.agrona.SystemUtil;
 
 import java.net.InetSocketAddress;
 
-import static io.aeron.driver.MinMulticastFlowControl.EMPTY_RECEIVERS;
-import static io.aeron.logbuffer.LogBufferDescriptor.computePosition;
 import static java.lang.System.getProperty;
 import static org.agrona.BitUtil.SIZE_OF_INT;
 import static org.agrona.BitUtil.SIZE_OF_LONG;
@@ -38,7 +34,7 @@ import static org.agrona.BitUtil.SIZE_OF_LONG;
  * Tracking of tagged receivers is done as long as they continue to send Status Messages. Once SMs stop, the receiver
  * tracking for that receiver will timeout after a given number of nanoseconds.
  */
-public class TaggedMulticastFlowControl implements FlowControl
+public class TaggedMulticastFlowControl extends AbstractMinMulticastFlowControl implements FlowControl
 {
     /**
      * URI param value to identify this {@link FlowControl} strategy.
@@ -58,9 +54,6 @@ public class TaggedMulticastFlowControl implements FlowControl
     public static final String PREFERRED_ASF = getProperty(PREFERRED_ASF_PROP_NAME, PREFERRED_ASF_DEFAULT);
     public static final byte[] PREFERRED_ASF_BYTES = BitUtil.fromHex(PREFERRED_ASF);
 
-    private volatile MinMulticastFlowControl.Receiver[] receivers = EMPTY_RECEIVERS;
-    private long receiverTimeoutNs;
-    private int groupMinSize;
     private long groupReceiverTag;
 
     /**
@@ -72,39 +65,12 @@ public class TaggedMulticastFlowControl implements FlowControl
         final int initialTermId,
         final int termBufferLength)
     {
-        receiverTimeoutNs = context.flowControlReceiverTimeoutNs();
-        groupReceiverTag = context.flowControlGroupReceiverTag();
-        groupMinSize = context.flowControlReceiverGroupMinSize();
+        receiverTimeoutNs(context.flowControlReceiverTimeoutNs());
+        receiverTag(context.flowControlGroupReceiverTag());
+        groupMinSize(context.flowControlReceiverGroupMinSize());
 
         final String fcValue = udpChannel.channelUri().get(CommonContext.FLOW_CONTROL_PARAM_NAME);
-
-        if (null != fcValue)
-        {
-            for (final String arg : fcValue.split(","))
-            {
-                if (arg.startsWith("t:"))
-                {
-                    receiverTimeoutNs = SystemUtil.parseDuration("fc receiver timeout", arg.substring(2));
-                }
-                else if (arg.startsWith("g:"))
-                {
-                    final int groupMinSizeIndex = arg.indexOf('/');
-
-                    if (2 != groupMinSizeIndex)
-                    {
-                        final int lengthToParse = -1 == groupMinSizeIndex ?
-                            arg.length() - 2 : groupMinSizeIndex - 2;
-                        groupReceiverTag = AsciiEncoding.parseLongAscii(arg, 2, lengthToParse);
-                    }
-
-                    if (-1 != groupMinSizeIndex)
-                    {
-                        groupMinSize = AsciiEncoding.parseIntAscii(
-                            arg, groupMinSizeIndex + 1, arg.length() - (groupMinSizeIndex + 1));
-                    }
-                }
-            }
-        }
+        FlowControlParameterParser.parse(fcValue, this::receiverTimeoutNs, this::groupMinSize, this::receiverTag);
     }
 
     /**
@@ -118,95 +84,8 @@ public class TaggedMulticastFlowControl implements FlowControl
         final int positionBitsToShift,
         final long timeNs)
     {
-        final long position = computePosition(
-            flyweight.consumptionTermId(),
-            flyweight.consumptionTermOffset(),
-            positionBitsToShift,
-            initialTermId);
-
-        final long windowLength = flyweight.receiverWindowLength();
-        final long receiverId = flyweight.receiverId();
         final boolean isTagged = isTagged(flyweight);
-        final long lastPositionPlusWindow = position + windowLength;
-        boolean isExisting = false;
-        long minPosition = Long.MAX_VALUE;
-
-        MinMulticastFlowControl.Receiver[] receivers = this.receivers;
-
-        for (final MinMulticastFlowControl.Receiver receiver : receivers)
-        {
-            if (isTagged && receiverId == receiver.receiverId)
-            {
-                receiver.lastPosition = Math.max(position, receiver.lastPosition);
-                receiver.lastPositionPlusWindow = lastPositionPlusWindow;
-                receiver.timeOfLastStatusMessageNs = timeNs;
-                isExisting = true;
-            }
-
-            minPosition = Math.min(minPosition, receiver.lastPositionPlusWindow);
-        }
-
-        if (isTagged && !isExisting)
-        {
-            final MinMulticastFlowControl.Receiver receiver = new MinMulticastFlowControl.Receiver(
-                position, lastPositionPlusWindow, timeNs, receiverId);
-            receivers = MinMulticastFlowControl.add(receivers, receiver);
-            this.receivers = receivers;
-            minPosition = Math.min(minPosition, lastPositionPlusWindow);
-        }
-
-        if (receivers.length < groupMinSize)
-        {
-            return senderLimit;
-        }
-        else if (receivers.length == 0)
-        {
-            return Math.max(senderLimit, lastPositionPlusWindow);
-        }
-        else
-        {
-            return Math.max(senderLimit, minPosition);
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public long onIdle(final long timeNs, final long senderLimit, final long senderPosition, final boolean isEos)
-    {
-        long minLimitPosition = Long.MAX_VALUE;
-        int removed = 0;
-        MinMulticastFlowControl.Receiver[] receivers = this.receivers;
-
-        for (int lastIndex = receivers.length - 1, i = lastIndex; i >= 0; i--)
-        {
-            final MinMulticastFlowControl.Receiver receiver = receivers[i];
-            if ((receiver.timeOfLastStatusMessageNs + receiverTimeoutNs) - timeNs < 0)
-            {
-                if (i != lastIndex)
-                {
-                    receivers[i] = receivers[lastIndex--];
-                }
-                removed++;
-            }
-            else
-            {
-                minLimitPosition = Math.min(minLimitPosition, receiver.lastPositionPlusWindow);
-            }
-        }
-
-        if (removed > 0)
-        {
-            receivers = MinMulticastFlowControl.truncateReceivers(receivers, removed);
-            this.receivers = receivers;
-        }
-
-        return receivers.length < groupMinSize || receivers.length == 0 ? senderLimit : minLimitPosition;
-    }
-
-    public boolean hasRequiredReceivers()
-    {
-        return receivers.length >= groupMinSize;
+        return handleStatusMessage(flyweight, senderLimit, initialTermId, positionBitsToShift, timeNs, isTagged);
     }
 
     private boolean isTagged(final StatusMessageFlyweight statusMessageFlyweight)
@@ -243,13 +122,8 @@ public class TaggedMulticastFlowControl implements FlowControl
         return groupReceiverTag;
     }
 
-    long receiverTimeoutNs()
+    void receiverTag(final long groupReceiverTag)
     {
-        return receiverTimeoutNs;
-    }
-
-    int groupMinSize()
-    {
-        return groupMinSize;
+        this.groupReceiverTag = groupReceiverTag;
     }
 }
